feat: part 1

+++ src/main/java/gregtechfoodoption/gui/widgets/KitchenRecipeWidget.java
@@ -289,7 +289,7 @@ public class KitchenRecipeWidget extends AbstractWidgetGroup implements IRecipeT
         tag.setTag("fluidInputs", fluidInputTag);
         tag.setTag("outputs", outputsTag);
         tag.setTag("fluidOutputs", fluidOutputsTag);
-        tag.setInteger("EUt", recipe.getEUt());
+        tag.setLong("EUt", recipe.getEUt());
 
         tag.setString("map", map);
         this.recipeWidget.deserializeNBT(tag);
